Refactor authentication and authorization concept
